+++ noteable_magics/planar_ally_client/errors.py
@@ -4,7 +4,7 @@ from typing import Any
 
 class PlanarAllyError(Exception):
     def __init__(self, msg):
-        super().__init__(msg or "There was an error while making that request, contact support.")
+        super().__init__(msg or "There was an error while making that request.")
 
     def user_error(self) -> str:
         return str(self)
@@ -16,7 +16,7 @@ class PlanarAllyAPIError(PlanarAllyError):
         self.body = body
         self.operation = operation
         super().__init__(
-            f"received {self.status_code} status from planar-ally for {self.operation}"
+            f"received {self.status_code} status from remove service for {self.operation}"
         )
 
     def user_error(self) -> str:
@@ -43,7 +43,7 @@ class PlanarAllyAPIError(PlanarAllyError):
 
 class PlanarAllyBadAPIResponseError(PlanarAllyError):
     def __init__(self):
-        super().__init__("Unable to parse response from remote service, contact support.")
+        super().__init__("Unable to parse response from remote service.")
 
 
 class PlanarAllyAPITimeoutError(PlanarAllyError):
@@ -52,5 +52,5 @@ class PlanarAllyAPITimeoutError(PlanarAllyError):
 
 
 class PlanarAllyUnableToConnectError(PlanarAllyError):
-    def __init__(self):
-        super().__init__("Unable to connect to planar-ally, contact support.")
+    def __init__(self, operation):
+        super().__init__(f"Unable to connect to remote service to perform operation '{operation}'")