+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.2] - 2021-07-27
+### Fixed
+- Explicitly use `sep=` kwarg for `pd.read_csv` to remove `FutureWarning`
+- Don't include dataframe index in SQL from `%create_or_replace_data_view` unless explicitly set with `--include-index`
+- Throw an error to stop cell execution when `NTBLMagic` fails
+
 ## [1.2.1] - 2021-07-19
 ### Fixed
 - Use `httpx` streams properly